+++ CHANGELOG.md
@@ -3,11 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.13.6
+## 3.4.13.7
 
 ### New Features
 
 ### Fixes
-- Issue with default tab selected on startup still referencing the old featured tab
 
 ### Misc